mod: refactored to make the upload form to be components
