+++ frontend/.vscode/settings.json
@@ -16,4 +16,10 @@
     }
   ],
   "prettier.eslintIntegration": true,
+  "python.testing.pytestArgs": [
+    "test"
+  ],
+  "python.testing.unittestEnabled": false,
+  "python.testing.nosetestsEnabled": false,
+  "python.testing.pytestEnabled": true,
 }
